+++ src/stream_futures.ts
@@ -63,7 +63,7 @@ interface TickerData {
   a: string;
   b: string;
   v: string;
-  Q: string;// last quantity traded 
+  Q: string; // last quantity traded
 }
 
 // Sort websocket data and store to redis
@@ -71,9 +71,9 @@ function handleTickerData(data: WebSocket.Data) {
   try {
     const tickers: TickerData[] = JSON.parse(data.toString());
     tickers.forEach((ticker) => {
-      const { s: symbol, c: price , Q:volume} = ticker;
+      const { s: symbol, c: price, Q: volume } = ticker;
 
-      //console.log("ticker", ticker);
+      console.log("futures ticker", ticker);
 
       // futures data is different than spot
 
@@ -83,7 +83,7 @@ function handleTickerData(data: WebSocket.Data) {
       if (Number(price) > 0 && price !== undefined) {
         let bid = price;
         let ask = price;
-        let vol = Number(price)* Number(volume);
+        let vol = Number(price) * Number(volume);
 
         const spr = Number(ask) - Number(bid);
         const sprPer = (Number(spr) / Number(ask)) * 100;
@@ -109,38 +109,6 @@ function handleTickerData(data: WebSocket.Data) {
     });
   } catch (error) {
     console.error("Error parsing message:", error);
-  }
-}
-
-// redis get key values and build an object
-export default async function fetchAllKeysAndValues() {
-  try {
-    // get all the keys in an array
-    const keys: string[] = await redisClient.keys("*");
-
-    //  console.log(`Total number of keys: ${keys.length}`);
-
-    // Array to store parsed key-value data
-    const parsedData: Array<Record<string, any>> = [];
-
-    // Fetch and parse each key's value
-    for (const key of keys) {
-      const value = await redisClient.get(key);
-
-      if (value) {
-        try {
-          // Parse the JSON string and add the object to the array
-          const jsonData = JSON.parse(value);
-          parsedData.push(jsonData);
-        } catch (error) {
-          console.error(`Error parsing value for key "${key}":`, error);
-        }
-      }
-    }
-
-    return parsedData;
-  } catch (error) {
-    console.error("Error fetching keys and values:", error);
   }
 }
